requirements.txt
Flask
python-dotenv
duckdb
python-binance
